Aspirepath-main/final/requirements.txt
streamlit==1.32.2
python-docx==1.1.0
PyYAML==6.0.1
bcrypt==4.1.2
pymongo==4.6.3
requests==2.31.0
